requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-adminlte-3==0.1.6
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-cuser==2017.3.16
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-rosetta==0.9.9
django-timezone-field==6.0.1
django-tinymce==3.6.1
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
kombu==5.3.2
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
polib==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.8
